Update to allow old vmu fs behavior.
New behavior for vmu reading provides the vmu
files without the headers readable by default.
In order to access the header data to manually
write it, `O_META` needs to be passed with the
file flags.

Additionally, `vmu_pkg_parse` now takes the
expected header size as a param for error checking.